week02/grpc-primes-solution/requirements.txt
grpcio>=1.66
grpcio-tools>=1.66
protobuf>=5.27
pytest>=8.3
